Add normalize and log1p transforms, using parametrize for tests (#1)
Adding total count normalization and log1p transforms. To save memory, transforms may change the original data matrix `X` that was used to construct an `AnnData` object by reference.

Tests now use `parametrize` which takes a list of functions to apply to arguments and re-runs each test with a different set of arguments.